Major changes:
1. Root rank now establishes a consensus on metrics all ranks have,
before triggering collectives (not doing so results in stalling and
incorrectness issues, as ranks may be participating in MPI_Reduce for
different metrics).

2. Metric aggregation is optimized as per the previous step. Since an
order of metrics is agreed upon now, they can be aggregated using arrays
rather than individual aggregation. THe older method still exists and
will be deleted in the following commit.